axfs_vfs

Macro impl_vfs_dir_default

macro_rules! impl_vfs_dir_default {
    () => { ... };
}

When implement VfsNodeOps on a directory node, add dummy file operations that just return an error.

This macro provides default implementations of file operations that return AxError::IsADirectory errors. It should be used when implementing VfsNodeOps for a directory node, as directories do not support file operations like read_at, write_at, fsync, and truncate.