axfs_vfs

Trait VfsOps

pub trait VfsOps: Send + Sync {
    // Required method
    fn root_dir(&self) -> VfsNodeRef;

    // Provided methods
    fn mount(&self, _path: &str, _mount_point: VfsNodeRef) -> VfsResult { ... }
    fn umount(&self) -> VfsResult { ... }
    fn format(&self) -> VfsResult { ... }
    fn statfs(&self) -> VfsResult<FileSystemInfo> { ... }
}

Filesystem operations.

This trait defines the operations that a filesystem must implement. All operations are safe by default and can be called concurrently from multiple threads (hence Send + Sync bounds).

Required Methods

fn root_dir(&self) -> VfsNodeRef

Get the root directory of the filesystem.

This method returns a reference to the root directory node of the filesystem. This is the only required method that must be implemented.

Returns

Returns a VfsNodeRef to the root directory.

Provided Methods

fn mount(&self, _path: &str, _mount_point: VfsNodeRef) -> VfsResult

Do something when the filesystem is mounted.

This method is called when the filesystem is mounted at a specific path. The default implementation does nothing.

Arguments

_path - The path where the filesystem is being mounted

_mount_point - A reference to the mount point directory node

Returns

Returns Ok(()) if the mount operation succeeds, or an error otherwise.

fn umount(&self) -> VfsResult

Do something when the filesystem is unmounted.

This method is called when the filesystem is unmounted. The default implementation does nothing.

Returns

Returns Ok(()) if the unmount operation succeeds, or an error otherwise.

fn format(&self) -> VfsResult

Format the filesystem.

This method formats the filesystem, erasing all existing data. The default implementation returns [AxError::Unsupported].

Returns

Returns Ok(()) if formatting succeeds, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the filesystem does not support formatting.

fn statfs(&self) -> VfsResult<FileSystemInfo>

Get the attributes of the filesystem.

This method retrieves information about the filesystem, such as total size, available space, etc. The default implementation returns [AxError::Unsupported].

Returns

Returns a FileSystemInfo containing filesystem attributes on success, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the filesystem does not support this operation.