axfs_vfs

Trait VfsNodeOps

pub trait VfsNodeOps: Send + Sync {

    // Provided methods
    fn open(&self) -> VfsResult { ... }
    fn release(&self) -> VfsResult { ... }
    fn get_attr(&self) -> VfsResult<VfsNodeAttr> { ... }
    fn read_at(&self, _offset: u64, _buf: &mut [u8]) -> VfsResult<usize> { ... }
    fn write_at(&self, _offset: u64, _buf: &[u8]) -> VfsResult<usize> { ... }
    fn fsync(&self) -> VfsResult { ... }
    fn truncate(&self, _size: u64) -> VfsResult { ... }
    fn parent(&self) -> Option<VfsNodeRef> { ... }
    fn lookup(self: Arc<Self>, _path: &str) -> VfsResult<VfsNodeRef> { ... }
    fn create(&self, _path: &str, _ty: VfsNodeType) -> VfsResult { ... }
    fn remove(&self, _path: &str) -> VfsResult { ... }
    fn read_dir(
        &self,
        _start_idx: usize,
        _dirents: &mut [VfsDirEntry],
    ) -> VfsResult<usize> { ... }
    fn rename(&self, _src_path: &str, _dst_path: &str) -> VfsResult { ... }
    fn as_any(&self) -> &dyn Any { ... }

}

Node (file/directory) operations.

This trait defines the operations that can be performed on a VFS node, which can be either a file or a directory. All operations are safe by default and can be called concurrently from multiple threads.

File Operations

The following methods are specific to files:

read_at - Read data from a file

write_at - Write data to a file

fsync - Synchronize file data to disk

truncate - Truncate a file

Directory Operations

The following methods are specific to directories:

parent - Get the parent directory

lookup - Look up a node by path

create - Create a new node

remove - Remove a node

read_dir - Read directory entries

rename - Rename or move a node

Provided Methods

fn open(&self) -> VfsResult

Do something when the node is opened.

This method is called when a node is opened for access. The default implementation does nothing.

Returns

Returns Ok(()) if the open operation succeeds, or an error otherwise.

fn release(&self) -> VfsResult

Do something when the node is closed.

This method is called when a node is closed after use. The default implementation does nothing.

Returns

Returns Ok(()) if the release operation succeeds, or an error otherwise.

fn get_attr(&self) -> VfsResult<VfsNodeAttr>

Get the attributes of the node.

This method retrieves metadata about the node, including its type, size, permissions, etc. The default implementation returns [AxError::Unsupported].

Returns

Returns a VfsNodeAttr containing the node’s attributes on success, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the node does not support this operation.

fn read_at(&self, _offset: u64, _buf: &mut [u8]) -> VfsResult<usize>

Read data from the file at the given offset.

This method reads up to buf.len() bytes from the file starting at offset. The actual number of bytes read is returned. The default implementation returns [AxError::InvalidInput].

Arguments

offset - The offset in the file to start reading from

buf - The buffer to read data into

Returns

Returns the number of bytes actually read on success, or an error otherwise.

Errors

Returns [AxError::InvalidInput] if called on a non-file node.

fn write_at(&self, _offset: u64, _buf: &[u8]) -> VfsResult<usize>

Write data to the file at the given offset.

This method writes up to buf.len() bytes to the file starting at offset. The actual number of bytes written is returned. The default implementation returns [AxError::InvalidInput].

Arguments

offset - The offset in the file to start writing to

buf - The buffer containing the data to write

Returns

Returns the number of bytes actually written on success, or an error otherwise.

Errors

Returns [AxError::InvalidInput] if called on a non-file node.

fn fsync(&self) -> VfsResult

Flush the file, synchronize the data to disk.

This method ensures that all data written to the file is persisted to the underlying storage device. The default implementation returns [AxError::InvalidInput].

Returns

Returns Ok(()) if synchronization succeeds, or an error otherwise.

Errors

Returns [AxError::InvalidInput] if called on a non-file node.

fn truncate(&self, _size: u64) -> VfsResult

Truncate the file to the given size.

If size is larger than the current file size, the file is extended with zeros. If size is smaller, the file is truncated. The default implementation returns [AxError::InvalidInput].

Arguments

size - The new size of the file in bytes

Returns

Returns Ok(()) if truncation succeeds, or an error otherwise.

Errors

Returns [AxError::InvalidInput] if called on a non-file node.

fn parent(&self) -> Option<VfsNodeRef>

Get the parent directory of this directory.

This method returns a reference to the parent directory of the current node. The default implementation returns None.

Returns

Returns Some(VfsNodeRef) if the node has a parent directory, or None if the node has no parent (e.g., it’s a file or root directory).

fn lookup(self: Arc<Self>, _path: &str) -> VfsResult<VfsNodeRef>

Lookup the node with given path in the directory.

This method searches for a node with the specified relative path within the directory. If found, it returns a reference to the node. The default implementation returns [AxError::Unsupported].

Arguments

path - The relative path to look up (e.g., “subdir/file.txt”)

Returns

Returns a VfsNodeRef to the found node, or an error if not found.

Errors

Returns [AxError::Unsupported] if the directory does not support lookup, or [AxError::NotFound] if the path does not exist.

fn create(&self, _path: &str, _ty: VfsNodeType) -> VfsResult

Create a new node with the given path in the directory.

This method creates a new file or directory with the specified path. If the node already exists, the method returns Ok(()). The default implementation returns [AxError::Unsupported].

Arguments

path - The path for the new node

ty - The type of node to create (file or directory)

Returns

Returns Ok(()) if the node was created or already exists, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the directory does not support creation.

fn remove(&self, _path: &str) -> VfsResult

Remove the node with the given path in the directory.

This method removes a file or directory at the specified path. The default implementation returns [AxError::Unsupported].

Arguments

path - The path of the node to remove

Returns

Returns Ok(()) if the node was removed successfully, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the directory does not support removal.

fn read_dir( &self, _start_idx: usize, _dirents: &mut [VfsDirEntry], ) -> VfsResult<usize>

Read directory entries into dirents, starting from start_idx.

This method reads directory entries (files and subdirectories) into the provided buffer, starting from the specified index. This allows for pagination of directory contents. The default implementation returns [AxError::Unsupported].

Arguments

start_idx - The starting index for reading entries (for pagination)

dirents - A mutable slice to store the directory entries

Returns

Returns the number of entries read on success, or an error otherwise.

Errors

Returns [AxError::Unsupported] if called on a non-directory node.

fn rename(&self, _src_path: &str, _dst_path: &str) -> VfsResult

Renames or moves existing file or directory.

This method renames or moves a node from src_path to dst_path. The operation can be within the same directory or across directories. The default implementation returns [AxError::Unsupported].

Arguments

src_path - The source path of the node to rename/move

dst_path - The destination path for the node

Returns

Returns Ok(()) if the rename/move operation succeeds, or an error otherwise.

Errors

Returns [AxError::Unsupported] if the directory does not support renaming.

fn as_any(&self) -> &dyn Any

Convert &self to &dyn Any that can use Any::downcast_ref.

This method enables downcasting to the concrete type that implements this trait. This is useful for accessing implementation-specific methods. The default implementation returns unimplemented!() and must be implemented by all concrete types.

Returns

Returns a reference to the node as &dyn Any.